backend/requirements.txt
# Backend dependencies for Reasoning Interface Generator
fastapi==0.116.1
uvicorn==0.32.1
pydantic==2.10.3
httpx==0.28.1
pypdf==5.4.0
python-docx==1.1.2
Pillow==11.1.0
